+++ dto/src/main/java/ca/on/oicr/gsi/runscanner/dto/ultima/CramAnalysisFile.java
@@ -1,6 +1,12 @@
 package ca.on.oicr.gsi.runscanner.dto.ultima;
 
 import ca.on.oicr.gsi.runscanner.dto.AnalysisFile;
+import ca.on.oicr.gsi.runscanner.dto.type.AnalysisFileFormat;
 
-// Represents one CRAM file produced by Ultima's initial CRAM generation workflow.
-public class CramAnalysisFile extends AnalysisFile {}
+// Represents one CRAM file produced by a CRAM generation workflow.
+public class CramAnalysisFile extends AnalysisFile {
+  @Override
+  public AnalysisFileFormat getFormatType() {
+    return AnalysisFileFormat.CRAM;
+  }
+}