+++ dto/src/test/java/ca/on/oicr/gsi/runscanner/rs/dto/test/dragen/JacksonTest.java
@@ -88,6 +88,7 @@ public class JacksonTest {
   }
 
   private static void assertAnalysisFileEqual(AnalysisFile one, AnalysisFile two) {
+    Assert.assertEquals(one.getFormatType(), two.getFormatType());
     Assert.assertEquals(one.getPath(), two.getPath());
     Assert.assertEquals(one.getCrc32Checksum(), two.getCrc32Checksum());
     Assert.assertEquals(one.getModifiedTime(), two.getModifiedTime());